fix: cleaned up PR a bit and fixed test failing
